Break out GitModule related methods
