+++ GitCommands/Git/GitCommandHelpers.cs
@@ -511,140 +511,6 @@ namespace GitCommands
             return args;
         }
 
-        /// <summary>
-        /// Parse the output from git-diff --name-status
-        /// </summary>
-        /// <param name="module">The Git module</param>
-        /// <param name="statusString">output from the git command</param>
-        /// <param name="staged">required to determine if <see cref="StagedStatus"/> allows stage/unstage.</param>
-        /// <returns>list with the parsed GitItemStatus</returns>
-        /// <seealso href="https://git-scm.com/docs/git-diff"/>
-        public static IReadOnlyList<GitItemStatus> GetDiffChangedFilesFromString(IGitModule module, string statusString, StagedStatus staged)
-        {
-            return new GetAllChangedFilesOutputParser(() => module).GetAllChangedFilesFromString_v1(statusString, true, staged);
-        }
-
-        /// <summary>
-        /// If possible, find if files in a diff are index or worktree
-        /// </summary>
-        /// <param name="firstId">from revision string</param>
-        /// <param name="secondId">to revision</param>
-        /// <param name="parentToSecond">The parent for the second revision</param>
-        /// <remarks>Git revisions are required to determine if <see cref="StagedStatus"/> allows stage/unstage.</remarks>
-        public static StagedStatus GetStagedStatus([CanBeNull] ObjectId firstId, [CanBeNull] ObjectId secondId, [CanBeNull] ObjectId parentToSecond)
-        {
-            StagedStatus staged;
-            if (firstId == ObjectId.IndexId && secondId == ObjectId.WorkTreeId)
-            {
-                staged = StagedStatus.WorkTree;
-            }
-            else if (firstId == parentToSecond && secondId == ObjectId.IndexId)
-            {
-                staged = StagedStatus.Index;
-            }
-            else if (firstId != null && !firstId.IsArtificial &&
-                     secondId != null && !secondId.IsArtificial)
-            {
-                // This cannot be a worktree/index file
-                staged = StagedStatus.None;
-            }
-            else
-            {
-                staged = StagedStatus.Unknown;
-            }
-
-            return staged;
-        }
-
-        public static IReadOnlyList<GitItemStatus> GetAssumeUnchangedFilesFromString(string lsString)
-        {
-            var result = new List<GitItemStatus>();
-            string[] lines = lsString.SplitLines();
-            foreach (string line in lines)
-            {
-                char statusCharacter = line[0];
-                if (char.IsUpper(statusCharacter))
-                {
-                    continue;
-                }
-
-                string fileName = line.SubstringAfter(' ');
-                GitItemStatus gitItemStatus = GitItemStatusConverter.FromStatusCharacter(StagedStatus.WorkTree, fileName, statusCharacter);
-                gitItemStatus.IsAssumeUnchanged = true;
-                result.Add(gitItemStatus);
-            }
-
-            return result;
-        }
-
-        public static IReadOnlyList<GitItemStatus> GetSkipWorktreeFilesFromString(string lsString)
-        {
-            var result = new List<GitItemStatus>();
-            string[] lines = lsString.SplitLines();
-            foreach (string line in lines)
-            {
-                char statusCharacter = line[0];
-
-                string fileName = line.SubstringAfter(' ');
-                GitItemStatus gitItemStatus = GitItemStatusConverter.FromStatusCharacter(StagedStatus.WorkTree, fileName, statusCharacter);
-                if (gitItemStatus.IsSkipWorktree)
-                {
-                    result.Add(gitItemStatus);
-                }
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Parse the output from 'git difftool --tool-help'
-        /// </summary>
-        /// <param name="output">The output string</param>
-        /// <returns>list with tool names</returns>
-        public static List<string> ParseCustomDiffMergeTool(string output)
-        {
-            var tools = new List<string>();
-
-            // Simple parsing of the textual output opposite to porcelain format
-            // https://github.com/git/git/blob/main/git-mergetool--lib.sh#L298
-            // An alternative is to parse "git config --get-regexp difftool'\..*\.cmd'" and see show_tool_names()
-
-            // The sections to parse in the text has a 'header', then break parsing at first non match
-
-            foreach (var l in output.Split('\n'))
-            {
-                if (l == "The following tools are valid, but not currently available:")
-                {
-                    // No more usable tools
-                    break;
-                }
-
-                if (!l.StartsWith("\t\t"))
-                {
-                    continue;
-                }
-
-                // two tabs, then toolname, cmd (if split in 3) in second
-                // cmd is unreliable for diff and not needed but could be used for mergetool special handling
-                string[] delimit = { " ", ".cmd" };
-                var tool = l.Substring(2).Split(delimit, 2, StringSplitOptions.None);
-                if (tool.Length == 0)
-                {
-                    continue;
-                }
-
-                // Ignore tools that must run in a terminal
-                string[] ignoredTools = { "vimdiff", "vimdiff2", "vimdiff3" };
-                var toolName = tool[0];
-                if (!string.IsNullOrWhiteSpace(toolName) && !tools.Contains(toolName) && !ignoredTools.Contains(toolName))
-                {
-                    tools.Add(toolName);
-                }
-            }
-
-            return tools.OrderBy(i => i).ToList();
-        }
-
         public static ArgumentString MergeBranchCmd(string branch, bool allowFastForward, bool squash, bool noCommit, string strategy, bool allowUnrelatedHistories, string mergeCommitFilePath, int? log)
         {
             return new GitArgumentBuilder("merge")
