+++ UnitTests/GitCommands.Tests/GitModuleTests.GetDiffChangedFilesFromString.ForScenario.Check_that_spaces_are_not_trimmed_in_file_names.approved.json
