+++ UnitTests/GitCommands.Tests/GitModuleTests.GetDiffChangedFilesFromString.ForScenario.Check_that_the_staged_status_is_None_if_not_IndexWorkTree2.approved.json
