+++ UnitTests/GitCommands.Tests/GitModuleTests.GetDiffChangedFilesFromString.ForScenario.Ignore_unmerged_in_conflict_if_revision_is_work_tree.approved.json
