+++ UnitTests/GitCommands.Tests/GitModuleTests.GetDiffChangedFilesFromString.ForScenario.Include_unmerged_in_conflict_if_revision_is_index.approved.json
