+++ UnitTests/GitCommands.Tests/GitModuleTests.GetDiffChangedFilesFromString.ForScenario.WorkTree1.approved.json
