+++ UnitTests/GitCommands.Tests/GitModuleTests.GetDiffChangedFilesFromString.ForScenario.WorkTree3.approved.json
